+++ scripts/pre-commit.sh
@@ -3,27 +3,27 @@
 set -eu
 
 test(){
-    bazelisk test //...
+    bazel test //...
 }
 
 fmt() {
-    bazelisk run :gazelle && bazelisk run :gofmt
+    bazel run :gazelle && bazel run @go_sdk//:bin/gofmt -- -s -w .
 }
 
 lint() {
-    bazelisk run :generate && bazelisk run :lint
+    bazel run @go_sdk//:bin/go -- generate -x ./... && bazel run @com_github_danmx_bazel_tools//golangci-lint:run -- run ./...
 }
 
 update_deps() {
-    bazelisk run :tidy && bazelisk run :gazelle -- update-repos -from_file=go.mod -to_macro=tools/repositories.bzl%go_repositories -prune=true
+    bazel run @go_sdk//:bin/go -- mod tidy && bazelisk run :gazelle -- update-repos -from_file=go.mod -to_macro=tools/repositories.bzl%go_repositories -prune=true
 }
 
 ci() {
-    bazelisk run :drone-fmt && bazelisk run :drone-sign
+    bazel run @com_github_danmx_bazel_tools//drone-cli:run -- fmt --save .drone.yml && bazel run @com_github_danmx_bazel_tools//drone-cli:run -- sign danmx/sigil --save
 }
 
 changelog(){
-    bazelisk run :changelog
+    bazel run @com_github_danmx_bazel_tools//git-chglog:run -- -o CHANGELOG.md
 }
 
 #shellcheck disable=SC2068