+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Chore
+- refactoring dev tools
 - **deps:** update dependency bazel_gazelle to v0.22.2
 - **deps:** update module aws/aws-sdk-go to v1.34.32
 - **deps:** update dependency io_bazel_rules_go to v0.24.3